+++ spec/genshin_data/characters_spec.rb
@@ -50,6 +50,14 @@ RSpec.describe GenshinData::Characters do
           expect(GenshinData.elements).to include(character.element)
         end
 
+        it "has a valid C3 bonus (#{character.constellation_three_bonus})" do
+          expect(%i[attack skill burst]).to include(character.constellation_three_bonus)
+        end
+
+        it "has a valid C5 bonus (#{character.constellation_five_bonus})" do
+          expect(%i[attack skill burst]).to include(character.constellation_five_bonus)
+        end
+
         it "has a valid weapon_type (#{character.weapon_type})" do
           expect(GenshinData.weapon_types).to include(character.weapon_type)
         end